Cleaned up comments and changed severity to Int
- Rewrote comments to be more useful for navigating the codebase
- Changed TrafficReport.severity from Float to Int since SeekBar returns Int

+++ app/src/main/java/dk/itu/moapd/x9/s25134/TrafficReport.kt
@@ -1,4 +1,3 @@
-// TrafficReport.kt
 package dk.itu.moapd.x9.s25134
 
 import android.os.Parcelable
@@ -8,5 +7,5 @@ import kotlinx.parcelize.Parcelize
 data class TrafficReport(
     val type: String,
     val description: String,
-    val severity: Float
+    val severity: Int
 ) : Parcelable
